requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-google-genai==0.0.6
langchain-community==0.0.10
google-generativeai==0.3.2
python-dotenv==1.0.0
reportlab==4.0.4
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
